Template refactor
