+++ dependency/src/main/java/org/wildfly/feature/pack/elytron/tls/dependency/Message.java
@@ -1,0 +1,20 @@
+package org.wildfly.feature.pack.elytron.tls.dependency;
+
+/**
+ * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
+ */
+public class Message {
+    private final String message;
+
+    public Message() {
+        this.message = null;
+    }
+
+    public Message(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return message;
+    }
+}
