--- dependency/src/main/java/org/wildfly/feature/pack/template/dependency/ExampleQualifier.java
@@ -1,17 +1,0 @@
-package org.wildfly.feature.pack.template.dependency;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.inject.Qualifier;
-
-/**
- * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
- */
-@Qualifier
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER })
-public @interface ExampleQualifier {
-}
