+++ example/pom.xml
@@ -19,15 +19,15 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>template-parent</artifactId>
-        <groupId>org.wildfly.extras.wildfly-feature-pack-template</groupId>
+        <groupId>org.wildfly.extras.elytron-tls</groupId>
+        <artifactId>elytron-tls-parent</artifactId>
         <version>1.0.0.Alpha-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>template-example</artifactId>
+    <artifactId>elytron-tls-example</artifactId>
     <packaging>war</packaging>
-    <name>WildFly Galleon FP Template - Example</name>
+    <name>Elytron TLS Subsystem - Example</name>
 
     <dependencies>
         <dependency>
@@ -64,7 +64,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>template-dependency</artifactId>
+            <artifactId>elytron-tls-dependency</artifactId>
             <scope>provided</scope>
         </dependency>
     </dependencies>
