--- feature-pack/src/main/resources/feature_groups/template-subsystem.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright 2019 Red Hat, Inc.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<!-- TODO Rename this file and the occurrences of 'template-subsystem' to what you set in TemplateExtension.SUBSYSTEM_NAME -->
-<feature-group-spec name="template-subsystem" xmlns="urn:jboss:galleon:feature-group:1.0">
-    <feature spec="subsystem.template-subsystem"/>
-</feature-group-spec>
