+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/elytron-tls-dependency/main/module.xml
@@ -1,0 +1,44 @@
+<!--
+  ~ Copyright 2019 Red Hat, Inc.
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~   http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+<module xmlns="urn:jboss:module:1.9" name="org.wildfly.elytron-tls-dependency">
+    <properties>
+        <property name="jboss.api" value="private"/>
+    </properties>
+
+    <resources>
+        <artifact name="${org.wildfly.extras.elytron-tls:elytron-tls-dependency}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-asn1}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-base}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-credential}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-keystore}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-provider-util}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-ssl-common}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-util}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-x500}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-x500-cert}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-x500-cert-acme}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-x500-cert-util}"/>
+        <artifact name="${org.wildfly.security:wildfly-elytron-x500-principal}"/>
+    </resources>
+
+    <dependencies>
+        <!-- We need these dependencies to trigger weld/CDI -->
+        <module name="javax.enterprise.api"/>
+        <module name="org.jboss.weld.api"/>
+        <module name="org.jboss.weld.core"/>
+        <module name="org.jboss.weld.spi"/>
+    </dependencies>
+</module>
