--- feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/extension/template-subsystem/main/module.xml
@@ -1,42 +1,0 @@
-<!--
-  ~ Copyright 2019 Red Hat, Inc.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<!-- TODO Rename this module to something matching your chosen subsystem name -->
-<module xmlns="urn:jboss:module:1.9" name="org.wildfly.extension.template-subsystem">
-    <properties>
-        <property name="jboss.api" value="private"/>
-    </properties>
-
-    <resources>
-        <!-- TODO Use the actual name of the subsystem artifact here -->
-        <artifact name="${org.wildfly.extras.wildfly-feature-pack-template:template-subsystem}"/>
-    </resources>
-
-    <dependencies>
-        <!-- Typical module dependencies for a subsystem -->
-        <module name="org.jboss.as.controller"/>
-        <module name="org.jboss.as.server"/>
-        <module name="org.jboss.logging"/>
-        <module name="org.jboss.modules"/>
-        <module name="org.jboss.staxmapper"/>
-        <module name="org.wildfly.security.manager"/>
-
-        <!--
-            Additional module dependencies. If they contain any classes that should be exposed to the user's code,
-            (e.g. in a deployment) make sure you add them in the subsystem's DependencyProcessor too
-        -->
-        <module name="org.wildfly.template-dependency"/>
-    </dependencies>
-</module>
