+++ feature-pack/wildfly-feature-pack-build.xml
@@ -14,10 +14,9 @@
   ~ limitations under the License.
   -->
 <!--
-TODO adjust the value of the 'producer' attribute to match the Maven groupId:artifactId of the maven project in ./pom.xml
 If you want to use a Galleon 'universe' (see the Galleon documentation) you could express that there instead.
 -->
-<build xmlns="urn:wildfly:feature-pack-build:3.1" producer="org.wildfly.extras.wildfly-feature-pack-template:template-feature-pack">
+<build xmlns="urn:wildfly:feature-pack-build:3.1" producer="org.wildfly.extras.elytron-tls:elytron-tls-feature-pack">
     <transitive>
         <dependency group-id="org.wildfly" artifact-id="wildfly-ee-galleon-pack">
             <name>org.wildfly:wildfly-ee-galleon-pack</name>
@@ -44,8 +43,7 @@ If you want to use a Galleon 'universe' (see the Galleon documentation) you coul
     <generate-feature-specs>
         <extensions>
             <standalone>
-                <!-- TODO Use the name of the module implementing your subsystem -->
-                <extension>org.wildfly.extension.template-subsystem</extension>
+                <extension>org.wildfly.extension.elytron-tls</extension>
             </standalone>
         </extensions>
     </generate-feature-specs>
