--- subsystem/src/main/java/org/wildfly/extension/feature/pack/template/subsystem/deployment/DependencyProcessor.java
@@ -1,67 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.wildfly.extension.feature.pack.template.subsystem.deployment;
-
-import org.jboss.as.server.deployment.Attachments;
-import org.jboss.as.server.deployment.DeploymentPhaseContext;
-import org.jboss.as.server.deployment.DeploymentUnit;
-import org.jboss.as.server.deployment.DeploymentUnitProcessor;
-import org.jboss.as.server.deployment.module.ModuleDependency;
-import org.jboss.as.server.deployment.module.ModuleSpecification;
-import org.jboss.modules.Module;
-import org.jboss.modules.ModuleLoader;
-
-/**
- * @author <a href="mailto:kabir.khan@jboss.com">Kabir Khan</a>
- */
-// TODO Rename this class to something which makes sense for your subsystem
-public class DependencyProcessor implements DeploymentUnitProcessor {
-
-    @Override
-    public void deploy(DeploymentPhaseContext phaseContext) {
-        DeploymentUnit deploymentUnit = phaseContext.getDeploymentUnit();
-
-        addModuleDependencies(deploymentUnit);
-    }
-
-    @Override
-    public void undeploy(DeploymentUnit context) {
-    }
-
-    private void addModuleDependencies(DeploymentUnit deploymentUnit) {
-        final ModuleSpecification moduleSpecification = deploymentUnit.getAttachment(Attachments.MODULE_SPECIFICATION);
-        final ModuleLoader moduleLoader = Module.getBootModuleLoader();
-
-        // Pull in dependencies needed by deployments in the subsystem
-
-        // This is needed if running with a security manager, and seems to be needed by arquillian in all cases
-        moduleSpecification.addSystemDependency(new ModuleDependency(moduleLoader, "org.wildfly.security.manager", false, false, true, false));
-        // TODO use the name of the modules after renaming, and add any other dependencies
-        // In this case we don't need any classes from the subsystem module itself so we don't need to add it to the
-        // deployment's module dependencies
-        // moduleSpecification.addSystemDependency(new ModuleDependency(moduleLoader, "org.wildfly.extension.template-subsystem", false, false, true, false));
-        moduleSpecification.addSystemDependency(
-                cdiDependency(new ModuleDependency(moduleLoader, "org.wildfly.template-dependency", false, false, true, false)));
-    }
-
-
-    private ModuleDependency cdiDependency(ModuleDependency moduleDependency) {
-        // This is needed following https://issues.redhat.com/browse/WFLY-13641 / https://github.com/wildfly/wildfly/pull/13406
-        moduleDependency.addImportFilter(s -> s.equals("META-INF"), true);
-        return moduleDependency;
-    }
-}
