+++ subsystem/src/main/resources/org/wildfly/extension/elytron/tls/subsystem/LocalDescriptions.properties
@@ -14,5 +14,7 @@
 # limitations under the License.
 #
 
-# TODO This should match the fully qualified name of the renamed TemplateExtension class
-org.wildfly.extension.elytron.tls.subsystem.ElytronTlsExtension
+# TODO rename these to the name you gave the subsystem in TemplateExtension
+elytron-tls=Elytron based subsystem for providing basic SSL/TLS features
+elytron-tls.add=Adds the Elytron TLS subsystem
+elytron-tls.remove=Removes the Elytron TLS subsystem
