--- subsystem/src/main/resources/org/wildfly/extension/feature/pack/template/subsystem/LocalDescriptions.properties
@@ -1,20 +1,0 @@
-#
-# Copyright 2019 Red Hat, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-# TODO rename these to the name you gave the subsystem in TemplateExtension
-template-subsystem=The Galleon Pack template subsystem
-template-subsystem.add=Adds the Galleon Pack template subsystem
-template-subsystem.remove=Removes the Galleon Pack template subsystem
