--- subsystem/src/main/resources/schema/template-subsystem_1_0.xsd
@@ -1,30 +1,0 @@
-<!--
-  ~ Copyright 2019 Red Hat, Inc.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~   http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-<!--
-    TODO Rename this file and the namespaces something that makes sense.
-    The namespaces should match those in SubsystemParser_1_0.NAMESPACE
--->
-<xs:schema xmlns:xs="http://www.w3.org/2001/XMLSchema"
-           targetNamespace="urn:wildfly:template-subsystem:1.0"
-           xmlns="urn:wildfly:template-subsystem:1.0"
-           elementFormDefault="qualified"
-           attributeFormDefault="unqualified"
-           version="1.0">
-
-    <!-- The subsystem root element -->
-    <xs:element name="subsystem">
-    </xs:element>
-</xs:schema>
