+++ testsuite/integration/subsystem/pom.xml
@@ -19,25 +19,25 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
-        <artifactId>template-testsuite-integration</artifactId>
-        <groupId>org.wildfly.extras.wildfly-feature-pack-template</groupId>
+        <artifactId>elytron-tls-testsuite-integration</artifactId>
+        <groupId>org.wildfly.extras.elytron-tls</groupId>
         <version>1.0.0.Alpha-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>template-testsuite-integration-subsystem</artifactId>
-    <name>WildFly Galleon FP Template - Testsuite: Subsystem</name>
+    <artifactId>elytron-tls-testsuite-integration-subsystem</artifactId>
+    <name>Elytron TLS Subsystem - Testsuite: Subsystem</name>
 
     <dependencies>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>template-feature-pack</artifactId>
+            <artifactId>elytron-tls-feature-pack</artifactId>
             <type>pom</type>
             <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>template-feature-pack</artifactId>
+            <artifactId>elytron-tls-feature-pack</artifactId>
             <type>zip</type>
             <exclusions>
                 <exclusion>
@@ -48,7 +48,7 @@
         </dependency>
         <dependency>
             <groupId>${project.groupId}</groupId>
-            <artifactId>template-dependency</artifactId>
+            <artifactId>elytron-tls-dependency</artifactId>
             <scope>test</scope>
         </dependency>
     </dependencies>
@@ -87,7 +87,7 @@
                                 </feature-pack>
                                 <feature-pack>
                                     <groupId>${project.groupId}</groupId>
-                                    <artifactId>template-feature-pack</artifactId>
+                                    <artifactId>elytron-tls-feature-pack</artifactId>
                                     <version>${project.version}</version>
                                     <inherit-packages>false</inherit-packages>
                                     <inherit-configs>false</inherit-configs>
@@ -105,8 +105,7 @@
                                         <layer>web-server</layer>
                                         <layer>jmx-remoting</layer>
                                         <!-- Layers from this FP  -->
-                                        <!-- TODO - rename to what you have called your layer -->
-                                        <layer>template-layer</layer>
+                                        <layer>elytron-tls-layer</layer>
                                     </layers>
                                 </config>
                             </configurations>
